requirements.txt
python-telegram-bot==13.7
SQLAlchemy==1.4.22
requests==2.25.1
